fix storing warm boot string in WarmBootHelper
Summary:
This fixes an issue where the warmBootDir_ field in
WarmBootHelper is a StringPiece and has no ownership over the
underlying memory. This changes it to a std::string so it copies
the string in to the class when it is attached.

Test Plan:
Canaried on rsw1jo.13.frc1, where I was seeing the issue and it
started warm booting.

Reviewed By: jasmeetbagga@fb.com

Subscribers: net-systems@

FB internal diff: D2000820

Tasks: 6796517

Signature: t1:2000820:1429247562:17eee7f8b274f18683a9b7a5486b04e5c43b8cb8