check both presence and dirty before reading any sfp data
Summary:
This fixes an issue where we detect sfp presence, but
fail to read the eeprom so the present_ flag and the dirty_ flag
are both true. LAter, when we try to publish stats and read this
SFP data, we crash because the data is dirty. This prevents
reading any data from the sfp when the dirty_ flag is set. We should
still track down why we fail to read the eeprom, but this prevents
throwing an error and crashing when we do.

Test Plan:
Reproed the issue by commenting out the call to setIdProm in
SFPModule::detectSFP. Canaried the change on bh with the comment still
in place and the issue went away

Reviewed By: adhipgupta@fb.com, klahey@fb.com

Subscribers: net-systems@

FB internal diff: D2004594

Tasks: 6809632

Signature: t1:2004594:1429408999:d64b8c6522b4a308562976ece32020bfd271b80d