refactored TunManager to implement StateObserver API
Summary: TunManager should also be a state observer so that it can be notified of state updates. This diff converts the api to use stateUpdated instead of startSync so it can be registered as a stateObserver. Note that the TunManager does not extend AutoRegisterStateObserver and handles registering and unregistering itself. The reason for this is that we only want to sync Tun interfaces after the initial config is applied, as interfaces probed from the OS would get blown away.

Test Plan: canaried on lab switch and passed unit tests. Will also test that changing the interfaces in the fboss config works between runs before landing.

Reviewed By: simpkins@fb.com

Subscribers: net-systems@

FB internal diff: D1967623

Tasks: 5789979, 6670185

Signature: t1:1967623:1429818098:306c2d69846d431d2c71fe6121d840cc4b48fd76