Only lock around notifying hardware in applyUpdate
Summary:
This makes our locking more granular during a state update and moves the notify observers call to after the hw stateChanged call. Previously, we locked around all of the work in applyUpdate, just to be safe. However, this leads to an issue when one of the observers needs to access the hwState. This happened in the arp expiration diff (D1682980) which had to be rolled back.

Also, since we no longer explicitly notify any members after D1967623 and all observers unregister themselves in the update thread, we are able to remove some of the locking in stop() that is used to prevent synchronization issues on shutdown. Previously, there could be issues when we start exiting and try to apply an update while partially exited, since we destroy the listeners before we stop the update/bg threads.

Depends on D1967623

Test Plan: Unit test, canary on lab switch

Reviewed By: simpkins@fb.com

Subscribers: net-systems@

FB internal diff: D1967668

Tasks: 5789979, 6809477

Signature: t1:1967668:1429817275:f5bc421eebb0b68c347a09c5fdb868dfe4df8a95