Make sure the vlan exists before trying to access arp/ndp tables in NeighborUpdater
Summary:
We shouldn't try to blindly access the vlan when a timeout expires
in NeighborUpdater. After D1967668, we notify after the state has been updated
so the time where a vlan is inconsistent can be much greater. There is an issue
right now on cold boots when we delete the placeholder vlan 1. All this code is
going to be replaced when D2004270 is landed, but we should fix this in the interim.

Test Plan:
Tested cold boot a number of times. Was seeing a crash here on cold
boots maybe 1/10 times before this fix.

Reviewed By: jasmeetbagga@fb.com

Subscribers: net-systems@

FB internal diff: D2027019

Signature: t1:2027019:1430182915:41bcd1f139007f6d5bcc120a8056819880fee53d