allow joint QSFP/SFP queries in thrift
Summary:
We want to be able to query both QSFP and SFP modules.  QSFPs have four
different channels, so we want to be able to retreive the per-channel
information independently.  For SFPs, we just return information for
a single channel.

This also adds support for cable length;  lengths are reported as
single-byte values, with a per-cable-type scaling factor which is
different on QSFPs and SFPs.  For instance, a single-mode length
is reported in kilometers, so we multiple by 1000 to report in meters.
A raw value of 255 indicates "greater in lenth than can be indicated;"
we report this as a negative value of the maximum size that can be
expressed by the interface, so -255,000 for a single-mode connection
that is longer than 255 km.

Test Plan:
Verify that it still builds, and that the fboss CLI still works:

fboss --hostname rsw1al.20.snc1.facebook.com port --detail

Reviewed By: simpkins@fb.com

Subscribers: net-systems@

FB internal diff: D2040047

Tasks: 6534083

Signature: t1:2040047:1430962305:436415b73451e03fdb81006a5eae98b64e5a6dd4