changed ThriftHandler ensureConfigured to check that we are not exiting
Summary:
This changes the ensureConfigured call in ThriftHandler to also
check if we are exiting and fail if so. This is not a full solution to
possible races between the ThriftHandler and shutdown, but is a definite
improvement.

Test Plan: unit test

Reviewed By: simpkins@fb.com

Subscribers: net-systems@

FB internal diff: D2047572

Tasks: 6862022

Signature: t1:2047572:1430882721:1b7933b18e00ab0f73a6378d1292a2fb277e797b