get rid of hwMutex_ in SwSwitch
Summary:
We already have a lock_ mutex in BcmSwitch that locks around
updates and some other calls in BcmSwitch. It looks like we were
unnecessarily acquiring both the hwMutex_ and this internal lock_ on a
number of calls (stateChanged, clearWarmBootCache...). This diff
removes the hwMutex_ from SwSwitch and adds calls to acquire lock_ in
the corresponding calls inside of BcmSwitch where necessary. Note that
this should be functionally equivalent to the locking scheme we had
before for calls to BcmSwitch from SwSwitch, except that the locking
occurs at the beginning of the BcmSwitch member function instead of
before the call to that function in SwSwitch. The locking could still
be a lot smarter, but this is a first step that just moves all the
locking into the BcmSwitch class

Test Plan: unit test + canary

Reviewed By: jasmeetbagga@fb.com

Subscribers: net-systems@

FB internal diff: D2047108

Tasks: 6862022

Signature: t1:2047108:1432072121:87f49315564e0bec11ff7033776c1346f9e4b291