Add thrift call in agent to force reload config file
Summary:
When we want to push new config to agent, currently we have to restart the agent. This diff introduces a thrift call which allows to load the config.

* It is also safer to do so because if the new config has some problems (invalid config), then this will let the agent run with older config.

Test Plan:
Currently, I have done the following.

* Run the agent on a test switch.
* Check the interfaces and details of one of the interface using
./_build/dbg/neteng/fboss/tools/cli/fboss.par
* Change the IP associated with one of the interfaces (in the config file) and do
./_build/dbg/neteng/fboss/tools/cli/fboss.par --hostname rsw1ai.20.snc1 --port 5909 reloadconfig
* Check the details of the changed interface and verify that IP address has changed.

* Also ran [fbconfig -r fboss && fbmake runtests].

I have not added any unit tests. I was not sure how to specify a config file in unit tests. If you think it is important to do so, we can discuss the details of that.

Reviewed By: jasmeetbagga@fb.com

Subscribers: net-systems@

FB internal diff: D2105171

Tasks: 7164534

Signature: t1:2105171:1432771981:456fba5b4878aaabb0dd031ade17dbda12afc3d9