Always sync tun interfaces after initial config is applied
Summary:
In D1967623 we made TunManager a state change observer.
However TunManager does not start listening to updates till after the
initial configuration is applied (this is done to get around the fact
that warm boot state does not have the tunnel intf information and we don't
want to cause the state update applied on warm boot to blow away the
tun interface). The problem though is that before D1967623 we always
synced tun interfaces after initial config was applied, but now we wait
for the next state update before syncing the tun interface. This is of
course problematic.
The reason that this happens only sometimes is that there is a race
here. Most often we would get a ARP, ndp update after the initial config
is applied and everything just works fine. In the racks that I saw the problem
we already had  arp, neighbor entries for the uplinks and all servers
in the rack were un provisioned. To further confirm this I deleted a
route via fboss_route tool and sure enough the front0 interface got configured.

Test Plan: Run it on a wedge switch

Reviewed By: simpkins@fb.com

Subscribers: net-systems@

FB internal diff: D2133666

Tasks: 7305971

Signature: t1:2133666:1433556288:9d1c5ece3e9a4e153f8e1b361bba0d0c882854f7