fix the race condition in using BlockingUpdateResult
Summary:
BlockingUpdateResult is used to allow thrift thread to wait for update done in
update thread. In our use case, BlockingUpdateResult object is allocated in the
stack of the thrift thread, and the pointer of the object is passed to the
update thread. Once the update is done, the update thread will signal thrift
thread using BlockingUpdateResult object.

The issue here is that once the thrift thread is signaled and the
BlockingUpdateResult object will be freed. Although
std::condition_variable::notify_one() is the last call in the update thread, we
still see blocking and sometimes crash in ::notify_one() call.

The diff uses shared_ptr<> for BlockingUpdateResult object to gurantee that the
object will still be alive in the case thrift thread is fast enough to finish
before notify_one() is completed.

Test Plan:
Developed a small program to simulate this situation. Was able to
reproduce notify_one() blocking within less than 1h of running. Changed the small program
to use shared_ptr<> and run the test for a long time, and confirm no thread is
blocking.

Reviewed By: aeckert@fb.com

Subscribers: net-systems@

FB internal diff: D2257974

Signature: t1:2257974:1437412703:0babb00b2b69a1f34723ecb2d64cc703aef5c3d4