drop log level for adding a route
Summary:
This drops the log level to 3 for adding a new route. When we
start up we spew log messages for each route added and it makes it
difficult to see other things going on. Based on our recommended log
levels it seems like 3 is the most appropriate level anyways.

Reviewed By: simpkins

Differential Revision: D2458297

fb-gh-sync-id: 9ee168e597858ff53656c65fd9e551a7edc9d320