Enhance getPortStatus call to expose transceiver info
Summary:
We want to be able to associate transceivers with ports.  It seemed
to make more sense to add this information the existing getPortStatus()
thrift call than create a new call (and more latency).  Presumably the
extra cycles to grab this information pale in comparison to the cost
of sending the data.

Reviewed By: simpkins

Differential Revision: D2366888

fb-gh-sync-id: f2455e7420ec24181d4e7fffb1a9ef6f3cce5340