Log the boot type, package info, revision info etc on startup.
Summary:
One of the more annoying things when debugging an issue on a
switch (especially when canarying new versions) is determining which
version was running when a crash occurred. This tries to resolve that
by logging the build_package_info and more build info to docstore on
startup. I am also going to write a tool that will answer this
question easily for us. Could have used scuba for this, but it seemed
like overkill for our needs.

Reviewed By: jasmeetbagga

Differential Revision: D2384287

fb-gh-sync-id: 0383ab0f48234796cdfabcd1d8905be6d8ced60e