Mark entries as pending on port down
Summary:
We want to tie adding egress entries back to ecmp entries
only once we can verify that ARP/NDP could be resolved for them
after port up. This lets us check that the port is ready to fwd
traffic after port up, before we turn the data hose on it. We
earlier had a 2 sec wait hack to get around this, but that just
delays the other updates and will be a pain if we have multiple
ports go down back to back.

So mark  ARP/NDP entries as pending on port down and reresolve them on
port up and only then update egress entries.

Reviewed By: aeckert

Differential Revision: D2539486

fb-gh-sync-id: fc605d53552f01ca8bc3fbd3eda6a5a5761ccaf9