Add support for changing LR4 QSFP power levels
Summary:
Wedge forces low power mode on QSFPs via a pin;  we need to set high
power mode on LR4 QSFPs via software.  After some experimentation, it
appears that we want to look at the Ethernet Compliance type of the
QSFP, and select the LR4s.  Checking the current power level doesn't
seem reliable -- some of the QSFPs don't appear to set it correctly.

I left in the code to log the characteristics of the QSFP, as it
appears that we'll be playing with this for some time, and with
other QSFPs, so it's probably useful to have this information,
and it's only retrieved when the QSFP is first detected.

Reviewed By: simpkins

Differential Revision: D2576236

fb-gh-sync-id: 12d1e92cd8b2680b3b319e8c1579beee627df9c4