Add support for drop routes
Summary:
There are cases where we need to program certain route
to have a drop action. This need showed up recently in the form of
a routing loop on 6 pack. The 6 pack announces summary routes from
FC101 for rsws. This summary is sent to SSWs. Now imagine if a rsw
drops out. Specific routes for it will be removed from 6 pack, but
the summary keeps getting announced to SSW (as expected). None of
the modules in 6 pack have this summary route in the h/w. When traffic
comes for this RSW subnet, it gets sent to 6 pack, which resolves it
to a SSW prefix which sends it back to 6 pack using the summary
route and a loop ensues.

I need to test this on wedges for /32, /128 host routes. Likely we
will have to make a small change there since the drop egress ID
does not need to be ref counted. I will send a separate diff if
changes are required for that.

Reviewed By: yogeshwer

Differential Revision: D2587534

fb-gh-sync-id: a004ac9856b41c439bbfc3a6820b5fe8811d00c9