Add FP rule to send packets to local IPs to high pri cosq
Summary:
Add FP rule to send packets to local IPs to high pri cosq.
We already prioritize packets that are directed to us, via
the nhop rx reason. However if these packets have a TTL<=1
the get classified as TTL<=1 packets even if they are destined
to us. TTL<=1 packets are put in the low pri queue, since these
are the packets that show up when routing loops occur.
The problem here was that EBGP by design sets TTL to 1 in non
multihop sessions. So in case of routing loops we ended up
dropping BGP packets.

Note I did not go for src or dst port == 179 as that would
catch BGP packets transiting through us.

Minor refactoring around FP initialization and creation.

Reviewed By: eastzone

Differential Revision: D2616667

fb-gh-sync-id: 53d82e3dc62bbe699ee755b526ac20d78c454055