Add support for configuring/unconfiguring static routes
Summary:
Look at the config and add/update any static routes.
On config reload do the same as above but also blow away any
static routes from previous config that did not show up in the
new config.

There are a couple of caveats to this support
- (minor) On warm boot if we have deleted routes in the new config
these won't be blown up till a sync fib happens or 30 seconds have
elapsed. This shouldn't be a problem, we retain stale bgp routes
for the same period after a warm boot too.
- We don't yet support admin distance for routes - t8910011

Reviewed By: petr

Differential Revision: D2596776

fb-gh-sync-id: f52f6ac99d2d61478a69fecb25b950e0522c203a