add a thrift call to transmit arbitrary packets
Summary:
Add thrift APIs to transmit arbitrary packets out front panel ports.  This adds
APIs for (1) sending directly to a specific port, (2) sending to a particular
VLAN, and letting the hardware pick which ports to send it out, (3) sending an
L3 payload, and letting the agent add the appropriate L2 next-hop header
information.

Reviewed By: jasmeetbagga

Differential Revision: D2596596

fb-gh-sync-id: ed64757ca1169a7029a76e7d1419af7e90b25564