Handle overflow of src_port when receiving packet from sdk
Summary:
Hacky workaround to an sdk bug where the 'src_port' field in the opennsl_pkt_t
is a signed 8-bit integer. This causes issues when the logical port # is > 127, as it
is for wedge100. This is likely fixed in the newest sdk, but we need this to work for
now. If we need to have more than 255 ports at some point we can change this back to
a 16 bit value.

Reviewed By: simpkins

Differential Revision: D2638185

fb-gh-sync-id: edf60d8d8d36790f71ab89ba4273ead25dafa6e4