Added logic to validate the configuration for a port group
Summary:
This diff adds logic to validate the configuration of the
ports in a port group is valid. This could eventually be used to
create a pre-update validation stage so we don't start applying a
state update we know will fail. For now it is just used to validate
the state during an update.

Reviewed By: simpkins

Differential Revision: D2567601

fb-gh-sync-id: f48a8803a249f6b30d7c22b72b7a1aba087d93d9