Add implementation for reconfiguring a port groups lane configuration
Summary:
This diff actually adds the logic to reconfigure a port group
to either SINGLE, DUAL or QUAD mode. There are a few changes in
this. First, I pushed most of the logic to enable, disable and program
a port into the BcmPort class. Then I add a step in stateChanged that
will reconfigure the port group if needed.

Note that all of this happens under the hood and the SwSwitch is not
changed at all. These checks do require examining the state as a
whole, which is a little awkward when we only have a StateDelta. In
the future we could decide to add some notion of dependent ports to
the SwitchState, so that it is more natural to apply updates that
require a reconfiguration.

Reviewed By: simpkins

Differential Revision: D2567726

fb-gh-sync-id: f46c7cec52fb74d871fc547eaa7a18b4e9d9e852