add unit tests for BcmPortGroup logic
Summary: This adds unit tests for calculating the configuration of a BcmPortGroup given the switch state. We can theoretically add more logic in here later. I had to factor out the actual calculation logic in to a static function since it is too difficult to mock the BcmSwitch and BcmPlatformPorts, which is slightly unfortunate. However at least now we have some unit tests for bcm code that can be run by fbmake.

Reviewed By: simpkins

Differential Revision: D2784943

fb-gh-sync-id: a19c63b72c940a591c9e075513b5db86c67f0afb