fix logic for enabling linkscan on warm boot
Summary:
My flexport diffs introduced a bug on warm boot because we
rely on BcmPort::enable being called to set up linkscan on
ports. However, enable is never called on warm boot because the ports
are already enabled. This changes the logic to set the linkscan mode
for ports during BcmPort::init, based on whether the port is enabled
or not in the ASIC.

Reviewed By: jasmeetbagga

Differential Revision: D2820722

fb-gh-sync-id: ad6bb4714bd79ac02d9d1f445107f572be5a1287