Move host routes to host table on Trident2
Summary:
First major commit towards moving host routes to host table.
The basic idea is on T2 for host routes, instead of creating
a route entry pointing to a egress object of nhops, for host routes
create another host object that points to the nhop(s) egress.

This is relatively straightforward, the only somewhat major change
now BcmHost no longer uniquely owns a corresponding object. This
is because for host routes, these will point to an existing
egress object for which host and egress objects already exists.
To accommodate this change start refcounting egress objects
and manage this via a table in BcmHostTable. I am still a bit
undecided about whether to abstract this it a separate BcmEgressTable
or manage these in BcmHostTable, but that is minor.

TODO:
a) Warm boot does not work across this change.
b) Besides a), need to get warm boot to work across earlier versions
where host routes were not in host table.

Reviewed By: yogeshwer

Differential Revision: D2350787

fb-gh-sync-id: b1a1def8a0a200b9e4eee7240678f5170418dba6