First diff towards updating ECMP objects on port state change
Summary:
The basic idea here is to track the egressIds that
go over a port, then on port state change either take them out
or add them back into Broadcom ECMP object. BcmEcmpEgress object
in addition now maintains the set of contained egressIds to
quickly determine whether we need to update it on port up/down
or not.

There are several TODO items here
a) We need to lock when calling BcmHost::linkStateChanged handler
since we would be reading some internal data structures which might
be in the process of getting updated. However acquiring this lock
leads to a deadlock at times. Need to debug why ?
b) Interaction of egress and ecmp egress objects with Neighbor
entries timing out needs to be figured out.
c) We need to be able to reconstruct the BcmSwitch state on
warm boot.

For a) I will debug and update this diff. For b) and c) new diffs
will come.

Reviewed By: yogeshwer

Differential Revision: D2454979

fb-gh-sync-id: 322bd9eee1f3b01c766c6ae60014c30a3139ece1