Read back ecmp->egress ids mapping and use that in link up/down handling
Summary:
Read back ecmp->egressIds mapping from warm boot files
to hold this map in BcmWarmBootCache. Then when constructing
egressIds-> Ecmp mapping use the above egress Ids rather
than the state from h/w since the latter would not have the downed ports,
while the BcmEcmpEgress that we went down with would have these.
To understand this consider a BcmEcmpEgress object X which has egress Is
A and B. Now imagine port for A goes down, we would then remove A from
h/w immediately but wait for a route update before updating X in memory.
If before that we shut down we need to revert back to the same state
we went down with (routes still thing X has A and B). Hence the above mapping.
With this mapping in place there are several cases to handle
(For cases lets consider the simple mapping of X -> {A, B} from above.
i) Simple case where no ports/egress Ids were removed from ECMP
entries and h/w and s/w were in sync. This case just works as before.
ii) A was removed from X and but routes synced and we got all the new routes
pointing to just B. Again this works just as before.
iii) A was removed but we shutdown before route sync. Now when we come
back up we will have a mapping of X -> {A, B} courtsey of the code above.
Now during sync fib we could get bgp's cached route (assuming bgp has not
noticed the peer loss and changed its routes yet) which will then pick
up X. Now we are back to the state we went down with and things proceed as
normal. If OTOH bgp does notice the change and updates its routes to
just point to B, we will get the new route and decrement reference to X.
This is again as things would have been had we never warm booted.
iv) A was removed, we warm booted but never heard of routes referring
to X in sync fib. X would be deleted post sync fib.
v) A was removed, we shut down, on warm boot we noticed that the
port came up before we ever heard of a route corresponding to X.
We would then use the warm boot cache to add A back to X. Since if
we don't and later we hear of routes pointing to X we won't ever add
A back since there will be no port up event to tell us to do so.
vi) Same as v) but now imagine that the port came up while controller
was not running (e.g. plug was reseated). At this point its unclear
whether we would ever get a link up event for this port. So to be
safe post warm boot we call link state change for all UP ports. Then
if this up port points to a egress entry that should be in a ECMP group
but is not, we add it.

There is one case missed here. If we warm boot very back to back, say
before a sync fib, then in this case BcmEcmpHost entry for X may not exist.
And we would not have written this to the warm boot file. So we would not
know the right mapping of Ecmp-> Egress Ids in s/w. This is quite a corner
case and only affects us when we shut down before sync fib (max 30 seconds
after start up), and have a actual link down event during this period.
I will file a separate task for this.

Reviewed By: yogeshwer

Differential Revision: D2473900

fb-gh-sync-id: ca3ba59a4ca17532aa1e1ad29666ea190a8c0249