Make port->egressIds mapping use a copy on write map
Summary:
This is a followup to D2483281, as outlined there and
in comments over BcmSwitch::linkStateChangedNoHwLock declaration,
we can't acquire BcmSwitch::lock_ in the link scan callback call chain
due to potential deadlocks. D2483281 got around this by calling
linkStateChanged from the update thread. However that means that these
calls can be potentially queued behind large updates causing a indeterminate
amount of delay before linkStateChanged gets called. We can't really afford
this delay on a link down event since we would keep losing packets till then.
So handle link down event in the link scan call back call chain, by using
only data structures that do not require a lock.
In these call backs we just need to read 2 data structures
a) Port->egressIds map to get the egressIds affected by the
port state change.
b) EgressIds for a ecmp object. This never changes after we have
created a Ecmp object.
To synchronize access for Port->egressId mapping maintain these
in a copy on write map, so the link scan thread can safely read
this while the update thread maybe doing updates on it. The updates
are always made on a cloned copy followed by a atomic update
of the port -> egressIDs map pointer.

For link up we have the opposite problem, adding egressIDs for the
newly up port too soon causes packet loss. Waiting 2 seconds (we
experimented with sub second, 1 second as well) after port came
up causes 0 packet loss (in both cases ARP/NDP entries were untouched
so loss was not due to ARP/NDP). We don't really want to wait in the
link scan call back, since that means if there a port down event
while we are waiting, we won't notice it until after the wait. So
put the port up events on the slow path by handling them in the update
thread and having the that update wait 2 seconds, before adding the
port back up. This is not a great solution. A better solution would
be to
a) Expire ARP/NDP entries for this port on port down
b) On port up, only add entries back after next hops have been arp/ndp
resolved.
The problem with that is that a) would direct all packets to CPU and
we don't yet have rate limiting for CPU traffic on all platforms. Once
we have that we can remove 2 seconds delay.

Reviewed By: yogeshwer

Differential Revision: D2488539

fb-gh-sync-id: 5c8101e44a09c777cb8d62272d1b1f60c8f4a8c2