Cleanup ecmp shrink/expand code and tackle some edge cases.
Summary:
Refactor some of the ECMP shrink/expand code
a) Make clear which functions can assume hw lock is acquired
or not.
b) Punt adding back egress entries to when ARP/NDP is resolved
c) When ARP/NDP in unresolved (for eg. on arp/ndp flush), update
ecmp entries as well. This is when port does not go down but arp/ndp
entries are expired/flushed.
d) During warm boot in addition to taking care of ports that came up while agent
was down, also take care of ports that went down.

Reviewed By: yogeshwer

Differential Revision: D2563866

fb-gh-sync-id: 76e2af09855b0cd80a04cdf860b506ee752fe5c3