Skip adding unresolved egress entries to ECMP group
Summary:
This is first step towards 2 tasks
a) t8837199 - Adding support for static routes for coalescing LC<->FC peerings
to be over loopback addresses (with static routes pointing to loopbacks).
With static routes in the picture we need to be careful when adding
unresolved nexthops to the route, since unlike bgp routes these routes
never go away, so the route won't be updated.
b) t8621401 - Tying next hop programming of routes to next hop ARP/NDP
being resolved. This diff tackles the ECMP portion (which is required
for a)).

Also removed a unused map data structure.

Reviewed By: petr

Differential Revision: D2580036

fb-gh-sync-id: 2129c5d611e425081193c6aa3d7f75d7ac72c706