Split wedge100 into its own platform
Summary:
This factors wedge100 in to its own platform instead of sharing the WedgePlatform
class and using the --mode flag. This will make it easier to add wedge100 specific code in
the future, as opposed to continually checking that flag. We could do a similar thing with
sixpack to keep the code cleaner.

Reviewed By: simpkins

Differential Revision: D2721978

fb-gh-sync-id: db53a1c241251c8088976cd5ba6b3d8b479e91f8
shipit-source-id: db53a1c241251c8088976cd5ba6b3d8b479e91f8