Use the correct I2C bus on wedge100
Summary:
On wedge100 we need to use the Wedge100I2CBus instead of the
WedgeI2CBus since the module selection logic is different. This
changes the platform to use the correct bus. This bug is why we were
resetting the CP2112 chip so often on wedge100 and (possibly) a fix
for transceiver presence command from the cli. Depends on D2721978

Reviewed By: simpkins

Differential Revision: D2853636

fb-gh-sync-id: 79b16f444e2f7b32bd423dbae6d45cd691e545f1
shipit-source-id: 79b16f444e2f7b32bd423dbae6d45cd691e545f1