Define new NeighborUpdater API -- DIFF 1/3
Summary:
This expands the NeighborUpdater API so that all updates to
neighbor entries now will go through the NeighborUpdater. Instead of
just starting the updater & listening for updates on the SwitchState,
the updater will now just listen for added + deleted vlans and will
take care of all of the needed state updates itself. It adds hooks for
various events that will affect the neighbor entry state. The updates
are performed by the ArpCache and NdpCache classes, which are just
stubs in this diff.

Reviewed By: simpkins

Differential Revision: D2004241

fb-gh-sync-id: 7149a76ca0fed64cf6bd9e11919b0aedbc86b3d1
shipit-source-id: 7149a76ca0fed64cf6bd9e11919b0aedbc86b3d1