added NeighborCache class to track sw neighbor state -- DIFF 2/3
Summary:
This adds the implementation for a pared down version of NDP.
This is encapsulated in the NeighborCache object, which stores
NeighborCacheEntries that do most of the heavy-lifting. These entries
correspond to NeighborEntry SwitchState nodes and store all the information
needed to program the corresponding SwitchState nodes, along with additional
info that is useful, but not needed to program the hardware. The actual logic
to program a NeighborCacheEntry is in the next diff.

Reviewed By: simpkins

Differential Revision: D2004266

fb-gh-sync-id: 5f6b8538a6bcdfbd21b618b678616ddd83c61dfc
shipit-source-id: 5f6b8538a6bcdfbd21b618b678616ddd83c61dfc