populate interfaceID in initial vlan map
Summary:
When we reconstruct the vlan map in BcmWarmBootCache, we do
not set the interfaceID field. This causes issues with my arp/ndp diffs
because we expect the interfaceID field to be populated when we create
the NeighborCaches.

This changes the code to also populate the interfaceID field when we
create the initial state. I also added a vlanChanged call to the
NeighborCaches so that it can handle interfaceID (and other) field
changes.

Depends on D2004270

Reviewed By: simpkins

Differential Revision: D2892352

fb-gh-sync-id: ec0e2be437ed50fb53bff07b70cfbd75c14c4d63
shipit-source-id: ec0e2be437ed50fb53bff07b70cfbd75c14c4d63