Bring internal code in sync with open source.
Summary:
Add in a missing script, so that D2832551 will apply.

NB: this commit should *not* be changed before landing -- it's to bring us in
sync with OSS so that external diffs will apply. This commit should be elided
from the OSS push, and any fixes/changes should happen in separate commits.

Reviewed By: simpkins

Differential Revision: D2887815

fb-gh-sync-id: e4b9ff2d4c8e8142b2f0f8e38fdd86acf282db54
shipit-source-id: e4b9ff2d4c8e8142b2f0f8e38fdd86acf282db54