fix port down callback to use NeighborCache
Summary:
Adam saw an issue while testing the weekly push where ports
would not become reachable after being flapped. This turned out to be
due to the logic that immediately sets neighbor entries to be pending
on port down, which had not been updated to use the NeighborCache,
which should be responsible for all updates to the NeighborEntries in
the switch state now. This fixes that issue and adds unit tests to
catch this issue in the future.

Reviewed By: simpkins

Differential Revision: D2904851

fb-gh-sync-id: edde1d83e0050449c45339a9b87d1ecc7048fcc0
shipit-source-id: edde1d83e0050449c45339a9b87d1ecc7048fcc0