add support for updates that disallow coalescing
Summary:
This adds a new type of update to the SwitchState-one that
does not allow coalescing with later updates. The reason for this is
that the BcmSwitch currently relies on seeing a pending neighbor entry
become valid for the ecmp expand logic. It would be possible for the
following updates to be coalesced together:

Entry A is REACHABLE
1. Update A to be PENDING
2. Update A to be REACHABLE

If these updates are combined together, the StateDelta we communicate
to hw would be a no-op for A, and the BcmSwitch would never add A back
to the relevant ecmp groups.

Differential Revision: D2912547

fb-gh-sync-id: dc19dbd240cbbe0c77b13346f4e220af3aca92b9
fbshipit-source-id: dc19dbd240cbbe0c77b13346f4e220af3aca92b9