fix stack smashing flushEntry call
Summary:
This fixes a bug in my arp/ndp diff that overwrites the stack
during an asynchronous flushEntry call. The bug is that we would set
flushed = true, where 'flushed' is captured by reference. This works
in the blocking case, but does not work when we call updateState
asynchronously. This leads to weird segfaults later on in the
code.This diff changes the blocking/non-blocking versions to use
different lambdas to avoid this bug.

Differential Revision: D2935127

fb-gh-sync-id: 9f145be7170b1a60b497785ac0112ab55999345e
fbshipit-source-id: 9f145be7170b1a60b497785ac0112ab55999345e