Ipv6 packets with hop limit == 1 that are destined to us are fine
Summary:
We saw that Arista SSWs are not consistent in what hop limit
they use in packets for BGP open messages to us. Sometimes they
pick 1 and at other times they pick 64. Agent code was dropping
v6 packets with hop limit set to 1 regardless of whether they
were destined to us or not. This does not seem right, since packets
destined to us don't need to go another hop to get to their destination.
so hop limit of 1 should be fine. We do the right thing for v4
packets with TTL == 1 already.

Differential Revision: D2969844

fb-gh-sync-id: 8427eaaa3bb90781d431c77cf8152c2113344c7d
fbshipit-source-id: 8427eaaa3bb90781d431c77cf8152c2113344c7d