fix transceiver mapping
Summary:
On wedge100 the mapping between qsfp number and logical port
number is not intuitive due to broadcom device port numbering
limitations. This fixes the logic for calculating logical port from
qsfp num/channel in the Wedge100Platform. This is pretty hacky, but
it is needed to get the qsfp presence detecting working correctly.

I plan to add a way to express this to the agent in the future so we
can get rid of this logic inside the platform.

Reviewed By: tfangit

Differential Revision: D2990499

fb-gh-sync-id: 19fce6e59f91a54fbbcb29d89366edc20685e8f6
fbshipit-source-id: 19fce6e59f91a54fbbcb29d89366edc20685e8f6