remove erroneous _DEPRECATED_ field
Summary:
It looks like D2239666 tried to add back in the supportedMTUs
field we used to have in the top level switch_config, but accidentally
added it as a member of the Port struct. We have been running for a long
time without this field and the supported MTU is not used anywhere (and
certainly shouldn't be on every port), so I am going to delete this.

Differential Revision: D3017310

fb-gh-sync-id: 3feb228e97cd1edb29d75fb1e6f2420b64c916cb
fbshipit-source-id: 3feb228e97cd1edb29d75fb1e6f2420b64c916cb