Add thrift api to set speeds of a set of ports. Add support in cli for same
Summary:
Overriding speeds temporarily is useful in custom
lab setups. Now with flexports support we can do this w/o
needing to hack the config, cold boot agent. Thrift api
takes a map of port->speed in Mbps as input and then
a) Checks if the speed update is legit
b) If so applies the update.

Invalid updates are handled nicely, in that, the thrift api
just throws a error for them, w/o touching the h/w.

Speeds can be any of the supported values from switch_cfg.thrit::PortSpeed
or -1, where -1 implies the port should be disabled. Disabling
ports is required when we want to run a few ports on a port
group at a higher speed in which they use multiple lanes (which
means no lanes maybe left for other ports in the same group, requiring
them to be disabled).

TODO: There is a bug where speed updates are not quite handling
the RX, TX pause setting correctly. Will fix that in a separate diff.

Reviewed By: simpkins

Differential Revision: D2996663

fb-gh-sync-id: 5a54a51a36f91f4f877215c85cf227416f572aa2
fbshipit-source-id: 5a54a51a36f91f4f877215c85cf227416f572aa2